Module projet_robert::configuration

Module de configuration

Tout ce qui permet de définir des paramêtres lors de sa compilation, se situe ici.

Constants

CANAL_NOM_DEFAUT
Nom du dictionnaire par défaut, créé par le programme et qui sert aussi de canal par défaut. Il ne peut et ne doit être jamais supprimé lors de l'exécution des requêtes des utilisateurs.

DEBUG
Définit sur le mode "débug" est actif (renvoi sur la console par défaut).

NBRE_MAX_CANAUX
Nbre maximum admissible de canaux dans le processus en cours.

NBRE_MAX_VALEURS
Nbre maximum admissible de valeurs pour chaque canal (dictionnaire).

PROFILS_PSEUDO_DEFAUT
Nom par défaut ('pseudo') d'un client TCP non-authentifié

PROFILS_SOURCE
Chemin vers le fichier des profils

TAILLE_LIGNE_MAX
Taille maximale admissible par ligne reçue sur un socket. Cette taille fournie donc la taille maximum admissible des requêtes pour le reste du programme.

TAILLE_TEXTE_MAX
Taille maximale admissible pour le texte contenu dans les dictionnaires.