Module projet_robert::resolution

Module de résolution des requêtes de clients

Modules

resoudre_administration
Sous-module de résolution "administration"

resoudre_canal
Sous-module de résolution "canal"

resoudre_numerique

resoudre_script

resoudre_texte

Structs

Retour
Structure définissant un 'Retour', afin d'uniformiser les messages à destination du client et l'état de résolution.

Enums

RetourType
Les retours peuvent être soit un texte statique (&'static str) - c'est-à-dire invariable et intégré au directement dans le code source du programme (efficacité), soit un texte généré par la fonction de résolution (String) - c'est-à-dire variable.

Functions

resoudre
Fonction de résolution centrale

resoudre_alterer
Fonction de résolution "altérer une valeur existante"

resoudre_definir
Fonction de résolution "définir une nouvelle valeur"

resoudre_lister
Fonction de résolution "lister toutes les valeurs d'un canal"

resoudre_obtenir
Fonction de résolution "obtenir une valeur existante"

resoudre_stop
Fonction de résolution : arrête la boucle principale du thread du client. Ne prend aucun argument (obligatoire).

resoudre_supprimer
Fonction de résolution "supprimer une valeur existante"

resoudre_tester
Fonction de résolution "tester l'existence d'un chemin"

Type Definitions

Resolveur
Un type spécifique au projet : le type 'Résolveur' est la signature d'une fonction de résolution, quelque soit le module de résolution. Elle prend deux paramètres : le contexte du socket ainsi qu'un objet permettant de récupèrer à la demande les arguments dits 'locaux' (propre à une requête). La fonction renvoie un objet "retour", qui sera transmis au client via une série d'octets écrite sur le socket. La définition de cette signature a pour principal but de soulager les signatures dans d'autres fonctions de résolution.