Function projet_robert::resolution::resoudre_administration::resoudre_eteindre

fn resoudre_eteindre(
    contexte: &mut Contexte, 
    arguments: ArgumentsLocaux
) -> Retour

Fonction de résolution locale "éteindre le programme"

Cette fonction doit être considérée comme avec un effet "prioritaire" : elle va annuler la boucle principale du service, ce qui provoquera l'arrêt de l'écoupe réseau ainsi que du programme. De plus, l'arrêt du service provoquera la fin de toutes les clients TCP actifs.

Une fois l'arrêt enclenché, il n'y a plus aucun moyen de revenir à l'état antérieur (perte des informations non-sauvegardées).