Module projet_robert::valeur

Module de définition des types de valeurs

Ce module permet de définition une énumération importante pour le programme : les types de valeurs possibles. Ces types sont "indépassables" au sein du processus, mais on peut les étendre avant la compilation, en ajoutant de nouveaux

Ces types peuvent être vus comme ceux fondamentaux dans un langage de programmation. Ils n'existe pas de notion d'alias cependant.

Un type peut parfois être "altérer" vers un autre. C'est-à-dire que le programme autorise des passages d'un type de valeur à l'autre - aux risques et périls du client, car certaines de ces altération sont possibles en perdant ou modifiant des informations.

L'altération depuis ou vers le type de valeur Objet est toujours impossible et interdit. Il n'a en effet aucun sens.

Dans le détail...

Aujourd'hui Robert compte 5 types de valeurs :

Les booléens (Booleen) qui ont un type fondamental interne bool. Les booléens peuvent représenter deux états ("vrai" / "1") ou ("faux" / "0").

Les relatifs (Relatif) qui ont un type fondamental interne i32 (extensible avant compilation à i64). Les booléens peuvent représenter des nombres entiers positifs ou négatifs.

Les flottants (Flottant) qui ont un type fondamental interne f32 (extensible avant compilation à f64). Les flottants peuvent représenter des nombres rationnels (avec la virgule), positifs ou négatifs. Attention, ils n'a aucune correction effectuée sur la précision des arrondis (voir 'cancellation' sur Wikipédia).

Les textes (Texte) qui ont un type interne String (référencables en &str en Rust). Seul le texte valide UTF-8 est reconnu pour du texte, comme d'ailleurs pour toutes les commandes.

Les objets (Objets) qui sont les regroupements de valeurs dans un dictionnaire (paire clé/valeurs, où la clé est utilisée sous forme de hash). Un objet est différent d'un canal, et sert à créer la récurcisité au sein des canaux ainsi que des représentations complexes d'informations.

A ce jour il manque deux types de valeurs appréciables : une chaîne binaire de taille arbitraire fixe ainsi que la représentation d'une date (UTC, probablement sous le format d'un timestamp 64bits). Ces formats seront peut-être ajoutés prochainement.

Enums

Valeurs